Replace TrustchainContext
