+++ tests/core/test_trustchain.py
@@ -8,7 +8,9 @@ from parsec.api.data import UserCertificate, DeviceCertificate, RevokedUserCerti
 from parsec.api.protocol import UserID, UserProfile
 from parsec.api.protocol.types import DeviceID
 from parsec.core.types import LocalDevice
-from parsec.core.trustchain import TrustchainContext, TrustchainError
+from parsec.core.trustchain import TrustchainError
+
+from libparsec.types import TrustchainContext
 
 
 class TrustchainData:
